Change value type. StatusEffect.Flexible.Actual.amount is int -> float

+++ Assets/Scripts/Battle/Damage/StatusEffector.cs
@@ -36,7 +36,7 @@ public static class StatusEffector
 				alreadyAppliedSameEffect.SetRemainStack(statusEffect.GetRemainStack());
 				if (statusEffect.IsOfType(StatusEffectType.Shield))
 					alreadyAppliedSameEffect.SetRemainAmount(
-						(int)(target.GetActualStat(statusEffect.GetAmountStat())*statusEffect.GetAmount()));
+						(target.GetActualStat(statusEffect.GetAmountStat())*statusEffect.GetAmount()));
 			}
 			else
 			{
@@ -44,7 +44,7 @@ public static class StatusEffector
 				if (statusEffect.IsOfType(StatusEffectType.Shield))
 				{
 					target.GetStatusEffectList()[target.GetStatusEffectList().Count].SetRemainAmount
-					((int)(target.GetActualStat(statusEffect.GetAmountStat())*statusEffect.GetAmount()));
+					((target.GetActualStat(statusEffect.GetAmountStat())*statusEffect.GetAmount()));
 				}
 			}
